20140908 Navigation entre les pages OK Navigation Menu/Settings OK Navigation Autres Pages/Settings pas OK Bouton html back to menu OK Bouton html back pas OK

+++ OneWebView_NavigateByStep/RefreshSettingViewController.m
@@ -59,7 +59,7 @@
             self.durationChoice = [[NSUserDefaults standardUserDefaults] objectForKey:@"durationChoice"];
         }
 
-        int iDuration = [self.durationValues indexOfObject:self.durationChoice];
+        long iDuration = [self.durationValues indexOfObject:self.durationChoice];
         
         [self.refresh selectRow:[self.intervalChoice integerValue]-1 inComponent:0 animated:YES];
         [self.refresh selectRow:iDuration inComponent:1 animated:YES];
